% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_cell_density.R
\name{anova_cell_density}
\alias{anova_cell_density}
\title{Boxplots of cell density.}
\usage{
anova_cell_density()
}
\description{
\code{anova_cell_density()} reproduces the compound figure of the cell
denisty data in section x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{anova_cell_density()} describes the cell density data located in
\code{./data-raw}.
}
\note{
Although this function is exported, \code{anova_cell_density()}
was not intended to be used outside of this package.
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Benjamin R. Gordon
}
