% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coralclass.R
\docType{package}
\name{coralclass}
\alias{coralclass}
\alias{coralclass-package}
\title{coralclass: A  package style research compendium for the coral
classification chapter of B. R. Gordon's PhD thesis.}
\description{
The package provides all the functions and data to reproduce the analysis
performed by Gordon et. al. This experiment examined the synergistic effects
of ocean warming and acidification on the coral Acropora aspera at Heron
Island in 2011.
}
\section{Functions}{

\itemize{
\item create_mzdata()
\item create_nmrdata()
\item figure_pca()
\item figure_tuning()
\item figure_vip()
\item mzdata_raw()
\item mzpls()
\item mzrf()
\item nmrpls()
\item nmrrf()
\item table_annotate()
\item table_cv_conmat()
\item table_cv_performance()
\item table_model_test()
}
}

\author{
Benjamin R. Gordon
}
