% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litmz.R
\name{create_litmz}
\alias{create_litmz}
\title{create litmz.}
\usage{
create_litmz(save.rda = FALSE, overwrite = FALSE)
}
\arguments{
\item{save.rda}{Logical indicating if a .rda file should be saved to
\code{/data}.}

\item{overwrite}{Logical indicating if existing files should be overwritten.}
}
\value{
Returns a dataframe of class \code{tbl_df}
}
\description{
\code{create_litmz()} cleans \code{./data-raw/metabs_literature.csv} for
future analysis
}
\details{
\code{create_litmz()} loads \code{metabs_literature.csv}, cleans up the
headers and creates \code{litmz.rda}.
}
\note{
\code{create_litmz()} was not intended to be used outside of this
package.
}
\author{
Benjamin R. Gordon
}
