% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mzdata.R
\name{create_mzdata}
\alias{create_mzdata}
\title{Create mzdata.}
\usage{
create_mzdata(parallel = FALSE, seed = 100, savecsv = FALSE,
  saverda = TRUE)
}
\arguments{
\item{parallel}{Logical indicating if missing values imputation should be run
in parallel. If \code{TRUE}, the default number of cores is equal to half the
available number of cores}

\item{seed}{An integer used for setting seeds of random number generation}

\item{savecsv}{Logical indicating if output should be saved as a \code{.csv}
file to the current working directory}

\item{saverda}{Logical indicating if a .rda file should be saved to /data}
}
\value{
Returns a dataframe of class tbl_df
}
\description{
\code{create_mzdata()} pre-processes the LCMS data used for modelling in
gordon01.
}
\details{
Initially, the function takes the raw output from xcms and removes unwanted
data (e.g. retention times, isotopes, peak counts etc.). Then, it creates
new categorical variables based on the sample information. Finally, it
replaces true non-detects with noise, removes poorly resolved mass features
and then replaces the small number of remining missing values using random
forest imputaion. The list below details the logic behind the missing values
imputation:
\itemize{
\item \strong{TRUE NON-DETECTS:}
 Replace values missing in one class, but not others, with a random number
 between zero and the minimum of the matrix (i.e. noise). To be considered a
 true non-detect, a class should be missing at least 60 precent of its
 values. Achieved with,
 \code{metabolomics::MissingValues(group.cutoff = 0.6)}
 \item \strong{POORLY RESOLVED MASS FEATURES:}
 Remove mass features with more than 90 percent missing values. Achieved
 with, \code{metabolomics::MissingValues(column.cutoff = 0.9)}
 \item \strong{FALSE NON DETECTS:}
 Remaining missing values will be computed using
 \code{missForest::missForest()}. Achieved with,
 \code{metabolomics::MissingValues(complete.matrix = FALSE)}
}
}
\note{
Using \code{parallel = TRUE} is not reproducible. Future versions of
this function may include support for reproducible RNG seeds when using
parallel processing. Although this function is exported,
\code{create_mzdata()} was not intended to be used outside of this package.
}
\seealso{
\code{\link[phdhelpr]{missing_values}}
\code{\link[doMC]{registerDoMC}}
\code{\link[missForest]{missForest}}
}
\author{
Benjamin R. Gordon
}
