% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrdata.R
\name{create_nmrdata}
\alias{create_nmrdata}
\title{create nmrdata.}
\usage{
create_nmrdata(savecsv = FALSE, saverda = TRUE, csvname = "nmrdata",
  remove.vars = TRUE, vars = c("3.31"), ...)
}
\arguments{
\item{savecsv}{Logical indicating if output should be saved as a \code{.csv}
file to the current working directory}

\item{saverda}{Logical indicating if a .rda file should be saved to /data}

\item{csvname}{The name of the output .csv file to be saved if TRUE}

\item{remove.vars}{Logical indicating if user-defined variables should be
removed}

\item{vars}{A character vector of variables to be removed}

\item{...}{Other arguments passed on to individual methods}
}
\value{
Returns a dataframe of class tbl_df
}
\description{
\code{create_nmrdata()} preprocesses the 1H-NMR data used for modelling
in gordon01.
}
\details{
\code{create_nmrdata()} takes the binned data from Bruker's Amix and creates
new categorical variables based on the sample ID's. It then removes any
unwanted variables as defined by the user (e.g. residual methanol or water
peaks) before saving the results to \code{./data}
}
\note{
\code{create_nmrdata()} was not intended to be used outside of this
package. Run this function with default values to reproduce the data used in
this package
}
\seealso{
\href{https://www.bruker.com/products/mr/nmr/nmr-software/nmr-software/amix/overview.html}{Bruker AMIX}
}
\author{
Benjamin R. Gordon
}
