% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pamdata.R
\name{create_pamdata}
\alias{create_pamdata}
\title{Create pamdata.}
\usage{
create_pamdata(path = "./data-raw/pamdata.csv", saverda = TRUE)
}
\arguments{
\item{path}{The path where the .csv is located}

\item{saverda}{Logical indicating if a .rda file should be saved to /data}
}
\value{
Returns a dataframe of class tbl_df
}
\description{
\code{create_pamdata()} reproduces the mean FvFm data.
}
\details{
This function reproduces the PAM data that was collected in this experiment.
}
\seealso{
\code{\link[tibble]{tibble}}
}
\author{
Benjamin R. Gordon
}
