% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_tables.R
\name{crossref_mzfeatures}
\alias{crossref_mzfeatures}
\title{Cross reference LCMS features}
\usage{
crossref_mzfeatures(ppm = 50, save.impvars = FALSE, save.matches = FALSE)
}
\arguments{
\item{ppm}{Numeric mass error to use for cross referencing}

\item{save.impvars}{Logical indicating if the table of the 20 most important
variables for each model should be saved to \code{./inst/extdata/}}

\item{save.matches}{Logical indication if the table of cross referencing
matches should be save to \code{./tables/}}
}
\value{
Returns a list containing the important and matched variables
}
\description{
\code{crossref_mzfeatures()} compares the 20 most important variables
identified in the PLS-DA and Random Forests models with those in
\code{\link{litmz}}.
}
\details{
\code{crossref_mzfeatures()} Loads the LCMS, PLS-DA and RF models and
creates a table of the 20 most important variables according to
\code{\link[caret]{varImp}}; three variables are calculated and bound to
this table:
\describe{
\item{$mz_neutral}{The neutral mass of the ion calculated by subtracting the
monoisotopic mass of the hydrogen ion (1.007276 Da)}
\item{$mz_low}{Equal to mz_neutral - the user defined ppm error}
\item{$mz_high}{Equal to mz_neutral + the user defined ppm error}
}
This table can be optionally saved to \code{./inst/extdata/} as a \code{.rds}
file. Finally, the important variables for each model are cross referenced
against a table of ions identified in the literature
(see \code{\link{litmz}}). Matches within the specified ppm error range are
retrieved and can be optionally saved to \code{./tables/} as a comma
separated \code{.txt} file.
}
\note{
Although this function is exported, \code{table_cv_conmat()} was not
intended to be used outside of this package.
}
\seealso{
\code{\link[caret]{varImp}}
\code{\link{litmz}}
}
\author{
Benjamin R. Gordon
}
