% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_tables.R
\name{cv_perform_table}
\alias{cv_perform_table}
\title{Cross validation performance table.}
\usage{
cv_perform_table(table.name = "CV_perform_table", save.table = FALSE)
}
\arguments{
\item{table.name}{Name of table.}

\item{save.table}{Logical indicating if the plot should be saved to
\code{./tables/}.}
}
\value{
returns a tibble with class \code{tbl_df}.
}
\description{
\code{cv_perform_table()} reproduces the cross validation performance table
in section x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{cv_perform_table()} loads the saved models in \code{./inst/extdata/};
calculates the cross validation confusion matrices and statistics for each
model; creates a table containing the accuracy and the class specific
sensitivity and specificity for each model; saves the table to
\code{./tables/} as a \code{.csv} file with a \code{.txt} file
extension.
}
\note{
The \code{.txt} file extension makes it easy to open the file in a
text editor and copy/paste the contents to Word, where the text can be easily
converted to a table. Although this function is exported,
\code{cv_perform_table()} was not intended to be used outside of this
package.
}
\seealso{
\code{\link[caret]{confusionMatrix}}
\code{\link[tibble]{tibble}}
}
\author{
Benjamin R. Gordon
}
