% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_cell_density.R
\name{figure_cell_density}
\alias{figure_cell_density}
\title{Boxplots of cell density.}
\usage{
figure_cell_density(view.plot = TRUE, save.plot = FALSE,
  plot.name = "cell_density_plot")
}
\arguments{
\item{view.plot}{Logical indicating if the plot should be printed to the plot
viewer.}

\item{save.plot}{Logical indicating if the plot should be saved to
\code{./figs/}.}

\item{plot.name}{Name of plot.}
}
\description{
\code{figure_cell_density()} reproduces the compound figure of the cell
denisty data in section x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{figure_cell_density()} summarises the cell density data located in
\code{./data-raw} The figure can be saved to \code{./figs/} if required.
}
\note{
Although this function is exported, \code{figure_cell_density()}
was not intended to be used outside of this package.
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Benjamin R. Gordon
}
