% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_fvfm.R
\name{figure_fvfm}
\alias{figure_fvfm}
\title{Plot of Photosynthetic Yield (FvFm).}
\usage{
figure_fvfm(save.plot = FALSE, plot.name = "fvfm_plot")
}
\arguments{
\item{save.plot}{Logical indicating if the plot should be saved to
\code{./figs/}.}

\item{plot.name}{Name of plot.}
}
\description{
\code{figure_fvfm()} produces the photosynthetic yield plot in section 4 of
Benjamin Gordon's PhD thesis.
}
\details{
\code{figure_fvfm()} plots the PAM data located in \code{./data}.
The figure can be saved to \code{./figs/} if required.
}
\note{
Although this function is exported, \code{figure_fvfm()}
was not intended to be used outside of this package.
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Benjamin R. Gordon
}
