% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_figures.R
\name{figure_misclass_temporal}
\alias{figure_misclass_temporal}
\title{Temporal misclassifications figure.}
\usage{
figure_misclass_temporal(view.plot = TRUE, save.plot = FALSE,
  plot.name = "temporal_misclass_figure")
}
\arguments{
\item{view.plot}{Logical indicating if the plot should be printed to the plot
viewer.}

\item{save.plot}{Logical indicating if the plot should be saved to
\code{./figs/}.}

\item{plot.name}{Name of plot.}
}
\description{
\code{figure_misclass_temporal()} reproduces the compound figure of the LCMS
and NMR model temporal misclassification results in section x.xx of Benjamin
Gordon's PhD thesis.
}
\details{
\code{figure_misclass_temporal()} loads the models that were saved to
\code{./inst/extdata/}; reproduces the training data; ascertains the temporal
information for each misclassification; summarises the misclassifications
according to the temporal data; creates side-by-side bar plots and places
them, with an appropriate plot title, into a single compound figure. The
figure can be saved to \code{./figs/} if required.
}
\note{
Although this function is exported, \code{figure_misclass_temporal()}
was not intended to be used outside of this package.
}
\seealso{
\code{\link[gridExtra]{gridExtra-package}}
\code{\link[ggplot2]{ggplot}}
\code{\link[grid]{grid-package}}
}
\author{
Benjamin R. Gordon
}
