% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_figures.R
\name{figure_tuning}
\alias{figure_tuning}
\title{Model tuning figure.}
\usage{
figure_tuning(view.plot = TRUE, save.plot = FALSE,
  plot.name = "tuning_figure")
}
\arguments{
\item{view.plot}{Logical indicating if the plot should be printed to the plot
viewer.}

\item{save.plot}{Logical indicating if the plot should be saved to
\code{./figs/}.}

\item{plot.name}{Name of plot.}
}
\description{
\code{figure_tuning()} reproduces the compound figure of the LCMS and NMR
model tuning results in section x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{figure_tuning()} loads the models and the saved gg objects in
\code{./inst/extdata/}; extracts the legend from the mzpca plot; adds plot
labels (a, b, c and d) and saves the plot to \code{./figs/} as a \code{.pdf}
file.
}
\note{
Although this function is exported, \code{figure_tuning()} was not
intended to be used outside of this package.
}
\seealso{
\code{\link[gridExtra]{gridExtra-package}}
\code{\link[ggplot2]{ggplot}}
\code{\link[grid]{grid-package}}
}
\author{
Benjamin R. Gordon
}
