% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_vip.R
\name{figure_vip}
\alias{figure_vip}
\title{Important variables figure.}
\usage{
figure_vip(number = 20, view.plot = TRUE, save.plot = FALSE,
  plot.name = "vip_plot")
}
\arguments{
\item{number}{The number of important variables to plot from each model}

\item{view.plot}{Logical indicating if the plot should be printed to the
viewer}

\item{save.plot}{Logical indicating if the plot should be saved to
\code{./figs/}.}

\item{plot.name}{Name of plot.}
}
\description{
\code{figure_vip()} was used to create a composite figure of the important
variables for each model
}
\details{
\code{figure_vip()} loads each model, determines the important variables
using \code{caret::varimp()} and creates a composite figure.
}
\note{
Although this function is exported, \code{figure_vip()} was not
intended to be used outside of this package.
}
\seealso{
\code{\link[caret]{varImp}}
\code{\link[ggplot2]{ggplot}}
}
\author{
Benjamin R. Gordon
}
