% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gordon01.R
\docType{package}
\name{gordon01}
\alias{gordon01}
\alias{gordon01-package}
\title{gordon01: A  package style research compendium for Gordon et. al. 2018}
\description{
The package provides all the functions and data to reproduce the analysis
performed by Gordon et. al. This experiment examined the synergistic effects
of ocean warming and acidification on the coral Acropora aspera at Heron
Island in 2011.
}
\section{Functions}{

\itemize{
\item create_litmz()
\item create_mzdata()
\item create_nmrdata()
\item figure_pca()
\item figure_tuning()
\item mzdata_raw()
\item mzpca()
\item mzpls()
\item mzrf()
\item nmrpca()
\item nmrpls()
\item nmrrf()
\item plot_base()
\item qual_colours()
\item sd_colours()
\item seq_colours()
\item table_cv_performance()
\item theme_brg_grid()
}
}

\author{
Benjamin R. Gordon
}
