% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litmz.R
\docType{data}
\name{litmz}
\alias{litmz}
\title{litmz}
\format{A tibble with 230 rows and 7 variables:
\describe{
  \item{reported_mz}{the m/z reported}
  \item{reported ion}{the ion reported}
  \item{monoiso_mass}{the monoisotopic mass}
  \item{molec_formula}{the molecular formula}
  \item{reported_name}{the reported name}
  \item{taxon}{the taxonomic information}
  \item{ref}{the reference where the information was reported. This is
  currently in custom endnote format and will be updated in the future.}
}}
\source{
Benjamin R. Gordon
}
\usage{
litmz
}
\description{
A dataset of published and/or known mz ions identified in either coral,
Symbiodinium or other closely related species.
}
\keyword{datasets}
