% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzdata.R
\docType{data}
\name{mzdata}
\alias{mzdata}
\title{Preprocessed LCMS data}
\format{A tibble with 101 rows and 1647 variables:
\describe{
  \item{sample_ids}{ID's for each sample}
  \item{class}{Treatment group. One of, control, eT (eleveated temperature),
  eCO2 (elevated CO2), eCO2eT (elevated CO2 and temperature), PBQC}
  \item{day}{Days of exposure to treatment}
  \item{tank}{Tank membership. Either L (left) or R (right)}
  \item{rep}{Sample replicate number}
  \item{class_day}{interaction variable between class and day variables}
}}
\source{
Benjamin R. Gordon
}
\usage{
mzdata
}
\description{
A dataset of positive ESI-LCMS mz variables for coral samples subjected to
experimental ocean warming and acidification at Heron Island in Feb 2011.
}
\keyword{datasets}
