% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzdata_raw.R
\name{mzdata_raw}
\alias{mzdata_raw}
\title{Feature detection using xcms.}
\usage{
mzdata_raw(saveoutput = FALSE, outputname = "example-mzdata-raw", ...)
}
\arguments{
\item{saveoutput}{Logical indicating if output should be saved}

\item{outputname}{The name of the output file to be saved if \code{TRUE}}

\item{...}{Other arguments passed on to individual methods}
}
\value{
Returns a dataframe of class \code{tbl_df}
}
\description{
Method used by Gordon et. al. to extract LCMS features from
\code{.mzXML} files using \code{xcms}.
}
\details{
This function was used to extract features from \code{.mzXML} data. The
function creates an \code{xcmsSet} object, groups peaks, corrects for
retention time drift, re-groups the peaks, fills missing peaks, annotates
adducts and isotopes, before saving the output to a \code{.csv} file in
\code{/data-raw}.
}
\note{
\code{mzdata_raw()} was not intended to be used outside of this
package. To keep the size of this package small, only a small example subset
of the raw LCMS \code{.mzXML} data is made available in
\code{./data-raw/example_mzxml_data/}. The full data set is available from
the package author.
}
\seealso{
\code{\link[xcms]{xcmsSet}}
\code{\link[CAMERA]{annotate}}
}
\author{
Benjamin R. Gordon
}
