% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzpca.R
\name{mzpca}
\alias{mzpca}
\title{Function to perform PCA for LCMS data.}
\usage{
mzpca(scale = FALSE, center = TRUE, view.plot = TRUE, save.pdf = FALSE,
  save.png = FALSE, save.gg = FALSE, plot.name = "mzpca_plot",
  seed = 1978, ...)
}
\arguments{
\item{scale}{Logical indicating if variables should be scaled to unit
variance.}

\item{center}{Logical indicating if variables should be zero centered.}

\item{view.plot}{Logical indicating if plot should be printed to the plot
viewer.}

\item{save.pdf}{Logical indicating if plot should be saved to \code{.pdf} in
the \code{./figs} directory.}

\item{save.png}{Logical indicating if plot should be saved to \code{.png} in
the \code{./figs} directory.}

\item{save.gg}{Logical indicating if ggplot object should be saved to
\code{./inst/extdata/mzpca_ggobject.rds}.}

\item{plot.name}{Name of plot.}

\item{seed}{An integer for setting the RNG state.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
returns a list with class \code{"prcomp"}.
}
\description{
\code{mzpca()} was used to perform the Principal Components Analysis of the
LCMS data (\code{./data/mzdata.rda})
}
\details{
\code{mzpca()} zero-centres the data and performs a PCA using singular value
decomposition according to \code{\link[stats]{prcomp}}. It outputs a
\code{prcomp} object and a plot, which can be saved as either a \code{.pdf}
or a \code{.png} file, or both.
}
\note{
Although this function is exported, \code{mzpca()} was not intended to
be used outside of this package.
}
\seealso{
\code{\link[stats]{prcomp}}
\code{\link[ggplot2]{ggplot}}
\code{\link{qual_colours}}
\code{\link{theme_brg_grid}}
}
\author{
Benjamin R. Gordon
}
