% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrdata.R
\docType{data}
\name{nmrdata}
\alias{nmrdata}
\title{Preprocessed NMR data}
\format{A tibble with 93 rows and 505 variables:
\describe{
  \item{sample_ids}{ID's for each sample}
  \item{class}{Treatment group. One of, control, eT (eleveated temperature),
  eCO2 (elevated CO2), eCO2eT (elevated CO2 and temperature), PBQC}
  \item{day}{Days of exposure to treatment}
  \item{tank}{Tank membership. Either L (left) or R (right)}
  \item{rep}{Sample replicate number}
  \item{class_day}{interaction variable between class and day variables}
}}
\source{
Benjamin R. Gordon
}
\usage{
nmrdata
}
\description{
A dataset of 1H-NMR variables (0.02 ppm wide bins) for coral samples
subjected to experimental ocean warming and acidification at Heron Island in
Feb 2011.
}
\keyword{datasets}
