% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pamdata.R
\docType{data}
\name{pamdata}
\alias{pamdata}
\title{Mean daily PAM data}
\format{A tibble with 28 rows and 3 variables:
\describe{
  \item{day}{day of measurement}
  \item{FvFm}{Photosynthetic yield}
  \item{class}{treatment class label}
}}
\source{
Benjamin R. Gordon
}
\usage{
pamdata
}
\description{
A dataset containing the mean PAM measurement for each class on days recorded
}
\keyword{datasets}
