% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.R
\name{plot_base}
\alias{plot_base}
\title{Custom base values for ggplot2}
\usage{
plot_base()
}
\description{
Custom base values for setting ggplot2 scales such as shape, color,
fill and xy axes. The code is mostly relevant to scatter plots.
}
\keyword{internal}
