% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{plot_shapes}
\alias{plot_shapes}
\title{Plot Shapes}
\usage{
plot_shapes(shape = NULL)
}
\arguments{
\item{shape}{The shape required e.g. \code{"triangle"}. \code{NULL} returns
all shapes.}
}
\value{
The default \code{NULL} returns a named numeric vector of
shapes 21:25.
}
\description{
Shapes used by plots in the gordon01 package.
}
\details{
The function \code{plot_shapes()} returns a vector of shape IDs for use in
\code{ggplot}. The shapes used are shapes 21:25 that have a \code{fill}
option
}
\note{
Although this function is exported, \code{plot_shapes()} was not
intended to be used outside of this package.
}
\seealso{
\href{http://sape.inf.usi.ch/quick-reference/ggplot2/shape}{Sape research group)}
}
\author{
Benjamin R. Gordon
}
