% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colours.R
\name{qual_colours}
\alias{qual_colours}
\title{Qualitative Colours}
\usage{
qual_colours(colour = NULL)
}
\arguments{
\item{colour}{The colour required e.g. \code{"orange"}. \code{NULL} returns
all colours.}
}
\value{
The default \code{NULL} returns a named character vector of all
colours available.
}
\description{
Custom qualitative, colour blind safe colours used by plots in the
gordon01 package.
}
\details{
The function \code{qual_colours()} returns the colour blind safe Okabe Ito
qualitative colour scale for distinguishing discrete groups.
}
\note{
Although this function is exported, \code{qual_colours()} was not
intended to be used outside of this package.
}
\seealso{
\href{http://serialmentor.com/dataviz/color-pitfalls.html#not-designing-for-color-vision-deficiency}{serialmentor.com}
and \href{http://jfly.iam.u-tokyo.ac.jp/color/}{Okabe and Ito (2008)}
}
\author{
Benjamin R. Gordon
}
