% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colours.R
\name{sd_colours}
\alias{sd_colours}
\title{Sarah and Duck Colours}
\usage{
sd_colours(colour = NULL)
}
\arguments{
\item{colour}{The colour required e.g. \code{"pink"}. \code{NULL} returns
all colours.}
}
\value{
The default \code{NULL} returns a named vector of all colours in the
palette.
}
\description{
Colours from Sarah and Duck.
}
\details{
The function \code{sd_colours()} returns a named vector of colours from the
Sarah and Duck cartoon.
}
\note{
Although this function is exported, \code{sd_colours()} was not
intended to be used outside of this package.
}
\examples{
sd_colours()

}
\seealso{
\href{https://www.sarahandduck.com}{sarahandduck.com}
}
\author{
Benjamin R. Gordon
}
