% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colours.R
\docType{data}
\name{seq_colours}
\alias{seq_colours}
\title{Yellow Orange Red Sequential colours}
\format{An object of class \code{character} of length 5.}
\usage{
seq_colours
}
\description{
Five sequential colours from the YlOrRd palette at colorbrewer.org.
}
\seealso{
\href{http://colorbrewer2.org/#type=sequential&scheme=YlOrRd&n=5}{colorbrewer.org}
}
\author{
Benjamin R. Gordon
}
\keyword{datasets}
