% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_tables.R
\name{table_cv_conmat}
\alias{table_cv_conmat}
\title{Cross validation confusion matrices.}
\usage{
table_cv_conmat()
}
\value{
returns a list of confusion matrices for each model.
}
\description{
\code{table_cv_conmat()} creates cross validation confusion matrices used in
table x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{table_cv_conmat()} loads the saved models in \code{./inst/extdata/};
reproduces then predicts the test data; creates a confusion matrix for each
model and returns them as a list of dataframes.
}
\note{
Although this function is exported, \code{table_cv_conmat()} was not
intended to be used outside of this package.
}
\seealso{
\code{\link[caret]{confusionMatrix}}
}
\author{
Benjamin R. Gordon
}
