% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_tables.R
\name{table_model_test}
\alias{table_model_test}
\title{Model testing performance table.}
\usage{
table_model_test(table.name = "model_testing_table", save.table = FALSE)
}
\arguments{
\item{table.name}{Name of table.}

\item{save.table}{Logical indicating if the plot should be saved to
\code{./tables/}.}
}
\value{
returns a tibble with class \code{tbl_df}.
}
\description{
\code{table_model_test()} reproduces the model testing performance table in
section x.xx of Benjamin Gordon's PhD thesis.
}
\details{
\code{table_model_test()} loads the saved models in
\code{./inst/extdata/}; reproduces then predicts the test data; calculates
the confusion matrices and statistics for the test data; creates a table
containing the accuracy and the class specific sensitivity and specificity
for each model; saves the table to \code{./tables/} as a \code{.csv} file
with a \code{.txt} file extension.
}
\note{
The \code{.txt} file extension makes it easy to open the file in a
text editor and copy/paste the contents to Word, where the text can be easily
converted to a table. Although this function is exported,
\code{table_model_test()} was not intended to be used outside of this
package.
}
\seealso{
\code{\link[caret]{confusionMatrix}}
\code{\link[tibble]{tibble}}
\code{\link[gordon01]{table_cv_performance}}
}
\author{
Benjamin R. Gordon
}
