% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg_grid.R
\name{theme_brg_grid}
\alias{theme_brg_grid}
\title{Custom theme for ggplot2}
\usage{
theme_brg_grid(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\description{
A custom \code{ggplot2} theme that includes major grid lines on both axes.
}
\details{
This theme is based on \code{\link[ggplot2]{theme_light}} with the
following major differences:
\itemize{
\item axis text to \code{size = 11}
\item axis title to \code{size = 14}
\item no panel border
\item no axis ticks
\item plot \code{margin = 11}
}
}
\note{
Although this function is exported, \code{theme_brg_grid()} was not
intended to be used outside of this package.
}
\seealso{
\code{\link[ggplot2]{theme_light}}
}
\author{
Benjamin R. Gordon
}
