% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gordon01.R
\docType{package}
\name{gordon01}
\alias{gordon01}
\alias{gordon01-package}
\title{gordon01: A  package style research compendium for Gordon et. al. 2018}
\description{
The package provides all the functions and data to reproduce the analysis
performed by Gordon et. al. This experiment examined the synergistic effects
of ocean warming and acidification on the coral Acropora aspera at Heron
Island in 2011.
}
\section{Functions}{

\itemize{
\item mzdata_raw
\item mzdata
}
}

\author{
Benjamin R. Gordon
}
