% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzpca.R
\name{mzpca}
\alias{mzpca}
\title{Function to perform PCA for LCMS data.}
\usage{
mzpca(scale = FALSE, center = TRUE, savepdf = TRUE, savepng = TRUE,
  plotname = "mzpca_plot", ...)
}
\arguments{
\item{scale}{Logical indicating if variables should be scaled to unit
variance.}

\item{center}{Logical indicating if variables should be zero centered.}

\item{savepdf}{Logical indicating if plot should be saved to \code{.pdf} in
the \code{./figs} directory.}

\item{savepng}{Logical indicating if plot should be saved to \code{.png} in
the \code{./figs} directory.}

\item{plotname}{Name of plot}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
returns a list with class \code{"prcomp"}.
}
\description{
\code{mzpca()} was used to perform the Principal Components Analysis of the
LCMS data (\code{./data/mzdata.rda})
}
\details{
\code{mzpca()} zero-centres the data and performs a PCA using singular value
decomposition according to \code{\link[stats]{prcomp}}. It outputs a
\code{prcomp} object and a plot, which can be saved as either a \code{.pdf}
or a \code{.png} file, or both.
}
\note{
Although this function is exported, \code{mzpca()} was not intended to
be used outside of this package. Run this function with default values to
reproduce the results in this package.
}
\seealso{
\code{\link[stats]{prcomp}}
\code{\link[ggplot2]{ggplot}}
\code{\link{plot_shapes}}
\code{\link{qual_colours}}
\code{\link{theme_brg_grid}}
}
\author{
Benjamin R. Gordon
}
