% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colours.R
\name{seq_colours}
\alias{seq_colours}
\title{Yellow Orange Red Sequential colours}
\usage{
seq_colours(colour = NULL)
}
\arguments{
\item{colour}{The colour required e.g. \code{"red"}. \code{NULL} returns
all colours.}
}
\value{
The default \code{NULL} returns a named vector of all colours in the
palette.
}
\description{
Five sequential colours from the YlOrRd palette at colorbrewer.org.
}
\details{
The function \code{()} returns a named vector of sequential colours from the
YlOrRd palette at colorbrewer.org.
}
\note{
Although this function is exported, \code{seq_colours()} was not
intended to be used outside of this package.
}
\examples{
seq_colours()

}
\seealso{
\href{http://colorbrewer2.org/#type=sequential&scheme=YlOrRd&n=5}{colorbrewer.org}
}
\author{
Benjamin R. Gordon
}
